Separate adapter and its implementation project
