+++ dev/com.ibm.ws.microprofile.opentracing.jaeger.adapter.impl/bnd.bnd
@@ -11,23 +11,24 @@
 -include= ~../cnf/resources/bnd/bundle.props,\
           ~../cnf/resources/bnd/app-resources.bnd
 
-bSymbolicName=com.ibm.ws.microprofile.opentracing.jaeger.adapter
+bSymbolicName=com.ibm.ws.microprofile.opentracing.jaeger.adapter.impl
 bVersion=1.0
 
 javac.source: 1.8
 javac.target: 1.8
 
-Bundle-Name: com.ibm.ws.microprofile.opentracing.jaeger.adapter
-Bundle-SymbolicName: com.ibm.ws.microprofile.opentracing.jaeger.adapter
+Bundle-Name: com.ibm.ws.microprofile.opentracing.jaeger.adapter.impl
+Bundle-SymbolicName: com.ibm.ws.microprofile.opentracing.jaeger.adapter.impl
 
-Bundle-Description:MicroProfile OpenTracing Jaeger Adapter Interface, version ${bVersion}
+Bundle-Description:MicroProfile OpenTracing Jaeger Adapter Interface Implementation, version ${bVersion}
 
 WS-TraceGroup: OPENTRACING
 
 Export-Package: \
-  com.ibm.ws.microprofile.opentracing.jaeger.adapter;version=1.0
+  com.ibm.ws.microprofile.opentracing.jaeger.adapter.impl;version=1.0
 
 Import-Package: \
+  !io.jaegertracing.*,\
   *
 
 Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
@@ -44,5 +45,7 @@ src: src
   com.ibm.websphere.org.osgi.service.component;version=latest, \
   com.ibm.wsspi.org.osgi.service.component.annotations;version=latest, \
   com.ibm.ws.org.osgi.annotation.versioning;version=latest, \
-  com.ibm.ws.io.opentracing.opentracing-api.0.31.0;version=latest
+  com.ibm.ws.io.opentracing.opentracing-api.0.31.0;version=latest, \
+  com.ibm.ws.microprofile.opentracing.jaeger.adapter;version=latest, \
+  io.jaegertracing:jaeger-core;version=0.34.0
