+++ dev/com.ibm.ws.microprofile.opentracing.jaeger.adapter.impl/src/com/ibm/ws/microprofile/opentracing/jaeger/adapter/impl/SenderConfigurationImpl.java
