+++ dev/com.ibm.ws.microprofile.opentracing.jaeger/.settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+encoding/bnd.bnd=UTF-8
