+++ dev/com.ibm.ws.microprofile.opentracing.jaeger/bnd.bnd
@@ -11,30 +11,34 @@
 -include= ~../cnf/resources/bnd/bundle.props,\
           ~../cnf/resources/bnd/app-resources.bnd
 
-bSymbolicName=com.ibm.ws.microprofile.opentracing.jaeger.adapter
+bSymbolicName=com.ibm.ws.microprofile.opentracing.jaeger
 bVersion=1.0
 
 javac.source: 1.8
 javac.target: 1.8
 
-Bundle-Name: com.ibm.ws.microprofile.opentracing.jaeger.adapter
-Bundle-SymbolicName: com.ibm.ws.microprofile.opentracing.jaeger.adapter
+Bundle-Name: com.ibm.ws.microprofile.opentracing.jaeger
+Bundle-SymbolicName: com.ibm.ws.microprofile.opentracing.jaeger
 
-Bundle-Description:MicroProfile OpenTracing Jaeger Adapter Interface, version ${bVersion}
+Bundle-Description:MicroProfile OpenTracing Jaeger Adapter, version ${bVersion}
 
 WS-TraceGroup: OPENTRACING
 
 Export-Package: \
-  com.ibm.ws.microprofile.opentracing.jaeger.adapter;version=1.0
+  com.ibm.ws.microprofile.opentracing.jaeger;version=1.0
 
 Import-Package: \
   *
+
+Private-Package: com.ibm.ws.microprofile.opentracing.jaeger.resources
 
 Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 
 instrument.disabled: true
 
-src: src
+src: src, resources
+
+-dsannotations: com.ibm.ws.microprofile.opentracing.jaeger.JaegerTracerFactory
 
 -buildpath: \
   com.ibm.ws.logging;version=latest, \
@@ -44,5 +48,9 @@ src: src
   com.ibm.websphere.org.osgi.service.component;version=latest, \
   com.ibm.wsspi.org.osgi.service.component.annotations;version=latest, \
   com.ibm.ws.org.osgi.annotation.versioning;version=latest, \
-  com.ibm.ws.io.opentracing.opentracing-api.0.31.0;version=latest
+  com.ibm.ws.opentracing.1.3;version=latest, \
+  com.ibm.ws.io.opentracing.opentracing-api.0.31.0;version=latest, \
+  com.ibm.ws.microprofile.opentracing.jaeger.adapter;version=latest, \
+  com.ibm.ws.microprofile.opentracing.jaeger.adapter.impl;version=latest, \
+  com.ibm.ws.crypto.passwordutil
